Move Matrix into its own public class
